[messages] Make dsme message type identifiers unique. Fixes JB#38500
The dsme message type identifiers are not unique:
  0x00000320 = SET_THERMAL_STATUS = BATTERY_EMPTY_IND
  0x00000506 = PROCESSWD_PING = PROCESSWD_PONG

While having the same id for PING and PONG happens not to be a logical
problem, nevertheless it is confusing and makes diagnostic logging
difficult without knowing the context.

Assign unique values for the duplicates.

Add how-to instructions in every header that defines message types to
make it less likely that similar problems arise again in the future.

Provide a library function for getting message type name so that
similar hacks can be removed from client code.

Add functions for accessing data in opaque dsme message objects.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>