Avoid unbounded strcpy() from environment to stack
If the "DSME_SOCKFILE" environment variable is set and has a
length longer than sizeof(c_addr.sun_path) (usually 108 bytes),
then strcpy() will overwrite values on the stack.

Avoid the issue by checking the string length before copying
it onto the stack, and returning an error if it's too long.

Note that getenv() isn't threadsafe, so this assumes that the
environment isn't changed, and it also assumes that the pointer
in dsmesock_default_location (as well as the pointed-to data)
isn't changed by another thread between the strlen() and the
strcpy(). A thread-safe version would most likely involve either
a copy of the string or the use of strncpy() (which might result
in a truncated path, which could lead to other bugs/crashes).

Signed-off-by: Thomas Perl <m@thp.io>