[ut_libdsme] Fix random test failures. Fixes JB#57276
When test application is executed as part of automated tests
there are failures at:

  Assertion 'strcmp(reason, "a-reason") == 0' failed

For one or another reason the failure can't be reproduced when
the test application is executed manually - which makes it difficult
to figure out what the problem is.

Add (controlled via command line options) diagnostic logging that can
hopefully be used to pinpoint causes of this/similar obscure problems.

Refactor starting/stopping of mock daemon child process so that
synchronization via SysV semaphore operations are not needed.

To keep behavior closer to the real thing, utilize wait for input or
timeout -logic instead of forcing sockets into blocking mode.

Remove linger control as it is not relevant for unix domain sockets.

Add more test points to both test_message and test_send_receive.

When sending dummy reply for test_send_receive, include terminating
nul character to payload - failure to do which was a potential cause
for the original problem (comparing unterminated text blob versus
regular terminated c-string yields results that depend on out of
bounds heap content).

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>